# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for the infrastructure archiving system
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /departements:
    get:
      operationId: getDepartements
      tags: [departements]
      summary: List all departments
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Departement"
    post:
      operationId: createDepartement
      tags: [departements]
      summary: Create a department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepartementBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Departement"
  /departements/{id}:
    get:
      operationId: getDepartement
      tags: [departements]
      summary: Get a department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Departement"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateDepartement
      tags: [departements]
      summary: Update a department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepartementBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Departement"
    delete:
      operationId: deleteDepartement
      tags: [departements]
      summary: Delete a department
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginBody:
      type: object
      properties:
        messager:
          type: string
        mot_pass:
          type: string
      required:
        - messager
        - mot_pass
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        nom:
          type: string
        prenom:
          type: string
        messager:
          type: string
        role:
          type: string
        grade:
          type: string
        adresse:
          type: string
        telephone_professional:
          type: ["integer", "null"]
        telephone_personnel:
          type: ["integer", "null"]
        is_chef_project:
          type: boolean
        id_departement:
          type: ["integer", "null"]
        nom_departement:
          type: ["string", "null"]
      required:
        - id
        - nom
        - prenom
        - messager
        - role
        - grade
        - adresse
        - is_chef_project
...
    Departement:
      type: object
      properties:
        id:
          type: integer
        nom:
          type: string
      required:
        - id
        - nom
    CreateDepartementBody:
      type: object
      properties:
        nom:
          type: string
      required:
        - nom
...